+++ com/password/prompt.js
@@ -1,0 +1,17 @@
+var url = require('url');
+
+exports = module.exports = function() {
+  
+  return function(req, res, next) {
+    var q = {};
+    if (res.locals.username) { q.username = res.locals.username; }
+    
+    return res.redirect(url.format({
+      pathname: '/login/password',
+      query: q
+    }));
+  };
+};
+
+exports['@implements'] = 'http://i.authnomicon.org/prompts/http/Prompt';
+exports['@name'] = 'password';